Moving new project to modules directory
